+++ cli_manager/commands/add_be_url.py
@@ -1,4 +1,4 @@
-""" The add-bind command. """
+""" The add-be-url command. """
 
 from json import dumps
 
@@ -18,17 +18,16 @@ def get_config_path(config_path=None):
     else:
         return os.path.join(os.path.dirname(get_script_path()), 'config.py')
 
-class AddDbEngine(Base):
-    """ Add DB Engine/Params to config """
+class AddBeUrl(Base):
+    """ Add Backend API URL to config """
 
     def run(self):
         config_path = get_config_path(self.options.get("--config-path"))
-
         config = config_editor.ConfigEditor(config_path)
-        bind_name = self.options.get("--bind-name") 
-        backend_version = self.options.get("--backend-version")
-        if bind_name and backend_version:
+        name = self.options.get("--name") 
+        value = self.options.get("--value")
+        if name and value:
             config.parse()
-            config.parsed.add_db_engine(bind_name, backend_version)
+            config.parsed.add_be_url(name, value)
             config.parsed_to_content()
             config.save()