+++ cli_manager/commands/set_redis_url.py
@@ -1,4 +1,4 @@
-""" The add-bind command. """
+""" The set-redis-url command. """
 
 from json import dumps
 
@@ -18,17 +18,14 @@ def get_config_path(config_path=None):
     else:
         return os.path.join(os.path.dirname(get_script_path()), 'config.py')
 
-class AddDbEngine(Base):
-    """ Add DB Engine/Params to config """
+class SetRedisUrl(Base):
+    """ Set Redis URL to config """
 
     def run(self):
         config_path = get_config_path(self.options.get("--config-path"))
-
         config = config_editor.ConfigEditor(config_path)
-        bind_name = self.options.get("--bind-name") 
-        backend_version = self.options.get("--backend-version")
-        if bind_name and backend_version:
-            config.parse()
-            config.parsed.add_db_engine(bind_name, backend_version)
-            config.parsed_to_content()
-            config.save()
+        value = self.options.get("--value")
+        config.parse()
+        config.parsed.set_redis_url(value)
+        config.parsed_to_content()
+        config.save()