+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Version 0.3.0 2018-Nov-26 22:57
+
+* Add new cli manager commands 
+* Add aux databases support
+* Add celery worker/beat units
+
 ## Version 0.1.0
 
 * Fix key_id value extraction in FullNodes helper model